登录b天骄返回username、发布温度，作业基本完成，但是数据库插入后是乱码

+++ KindgartenPlatform/src/action/mobile/BasicAction.java
@@ -88,7 +88,7 @@ public class BasicAction extends ActionSupport {
 					temp.setTeacherId(Long.valueOf(teacherId));
 				}
 				if (temperature != null && !temperature.isEmpty()) {
-					temp.setTemperature(Integer.valueOf(temperature));
+					temp.setTemperature(Float.valueOf(temperature));
 				}
 				if (childId != null && !childId.isEmpty()) {
 					temp.setChildId(Long.valueOf(childId));
